+++ components/Error/Error.tsx
@@ -1,10 +1,19 @@
 import danger from 'appdeptus/assets/lotties/danger.json'
 
 import LottieView from 'lottie-react-native'
+import { type ComponentProps } from 'react'
+import Button from '../Button'
 import Text from '../Text'
 import { VStack } from '../ui'
 
-const Error = () => (
+type ErrorProps = {
+  button?: ComponentProps<typeof Button>
+  caption?: string
+  description?: string
+  title?: string
+}
+
+const Error = ({ button, caption, description, title }: ErrorProps) => (
   <VStack className='flex-1 items-center justify-center'>
     <LottieView
       autoPlay
@@ -14,17 +23,27 @@ const Error = () => (
       }}
       source={danger}
     />
-    <VStack>
+    <VStack space='md'>
       <Text
         className='text-center'
         family='heading-regular'
         size='2xl'
       >
-        Astropathic Miscommunication!
+        {title ?? 'Astropathic Miscommunication!'}
       </Text>
+      {caption ? (
+        <Text
+          className='text-center'
+          family='body-bold'
+          size='lg'
+        >
+          {caption}
+        </Text>
+      ) : null}
       <Text className='text-center'>
-        You may want to pull to retry a new one!
+        {description ?? 'You may want to pull down to retry a new one!'}
       </Text>
+      {button ? <Button {...button} /> : null}
     </VStack>
   </VStack>
 )